net/usrsock: increase the send/recv() length limit to UINT32_MAX
VELAPLATFO-1058

change request type to uint32_t to the impove the throughput

+++ net/usrsock/usrsock_recvmsg.c
@@ -174,9 +174,9 @@ static int do_recvfrom_request(FAR struct usrsock_conn_s *conn,
       addrlen = UINT16_MAX;
     }
 
-  if (buflen > UINT16_MAX)
+  if (buflen > UINT32_MAX)
     {
-      buflen = UINT16_MAX;
+      buflen = UINT32_MAX;
     }
 
   /* Prepare request for daemon to read. */
