+++ net/usrsock/usrsock_sendmsg.c
@@ -174,9 +174,9 @@ static int do_sendto_request(FAR struct usrsock_conn_s *conn,
       req.buflen += msg->msg_iov[i].iov_len;
     }
 
-  if (req.buflen > UINT16_MAX)
+  if (req.buflen > UINT32_MAX)
     {
-      req.buflen = UINT16_MAX;
+      req.buflen = UINT32_MAX;
     }
 
   bufs[0].iov_base = (FAR void *)&req;
